requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy
python-dotenv
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
vllm
langchain
faiss-cpu
pypdf
sentence-transformers